fix(filters): Adds filters functionalities
- Keeps the data after the filters config is updated

- Keeps data after filters config updates

The code is reorganized, the public API methods are grouped now

The control focus is kept event after the filters config is updated and the component is re-rendered

Creates a filters master map which store data such as:
  - Filter configuration
  - Filter control reference
It can be extended later to include more relevant data

- Adds function to adapt a filter's value

Before sending the o-filter-change event, the filter's value
is adapted (converted) to a proper expected value. i.e.
an empty string, array or object is converted to undefined

- Adds show/hide filter

- Adds get filter config/control by name